+++ utils/changelog.md
@@ -2,6 +2,11 @@
 
 ## WIP
 
+- Add `Vulkan.Utils.Requirements`, a module with several helpers for the
+  `Requirement` types found in `Vulkan.Requirement`
+- Change the creation helpers in `Vulkan.Utils.Initialization` to use the new
+  `Requirement` types.
+
 ## [0.3] - 2020-11-24
 
 - Change type of pickPhysicalDevice to return Nothing instead of throwing